Styled body and progress container
I began by adding styling to the body of my code by adding a background color that is pale yellow in color and aligning everything to the center. I also aligned the text to the center using the container selector. Lastly, I also used the progress-conatiner selector to add space in between the numbers on the screen and making sure they were spread out over the entirety of the screen.