Started creating progress line
I began styling the progress line I want to be shown as the user presses the "next" button. I used the pseudoselector "before" to have a starting point where the line is just grey. I want the line to be a light shade of blue and it would transition smoothly with the "ease" transition when the line progresses.