button stlying
I styled the buttons to have a text color of white, padding of 6px on top and bottom and 20px on the left and right, I also made the font size a little larger and the text inherited the font Ive been using throughout this project (roboto, sans-serif). I added a transform property to the buttons so that they shift down in size when they are clicked.